--- src/components/table/table.component.jsx
@@ -1,188 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import {makeStyles, useTheme} from '@material-ui/core/styles';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableFooter from '@material-ui/core/TableFooter';
-import TablePagination from '@material-ui/core/TablePagination';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-import IconButton from '@material-ui/core/IconButton';
-import FirstPageIcon from '@material-ui/icons/FirstPage';
-import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
-import KeyboardArrowRight from '@material-ui/icons/KeyboardArrowRight';
-import LastPageIcon from '@material-ui/icons/LastPage';
-import TableHead from '@material-ui/core/TableHead';
-
-const useStyles1 = makeStyles((theme) => ({
-    root: {
-        flexShrink: 0,
-        marginLeft: theme.spacing(2.5),
-    },
-}));
-
-function TablePaginationActions(props) {
-    const classes = useStyles1();
-    const theme = useTheme();
-    const {count, page, rowsPerPage, onChangePage} = props;
-
-    const handleFirstPageButtonClick = (event) => {
-        onChangePage(event, 0);
-    };
-
-    const handleBackButtonClick = (event) => {
-        onChangePage(event, page - 1);
-    };
-
-    const handleNextButtonClick = (event) => {
-        onChangePage(event, page + 1);
-    };
-
-    const handleLastPageButtonClick = (event) => {
-        onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-    };
-
-    return (
-        <div className={classes.root}>
-            <IconButton
-                onClick={handleFirstPageButtonClick}
-                disabled={page === 0}
-                aria-label="first page"
-            >
-                {theme.direction === 'rtl' ? <LastPageIcon/> : <FirstPageIcon/>}
-            </IconButton>
-            <IconButton onClick={handleBackButtonClick} disabled={page === 0} aria-label="previous page">
-                {theme.direction === 'rtl' ? <KeyboardArrowRight/> : <KeyboardArrowLeft/>}
-            </IconButton>
-            <IconButton
-                onClick={handleNextButtonClick}
-                disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                aria-label="next page"
-            >
-                {theme.direction === 'rtl' ? <KeyboardArrowLeft/> : <KeyboardArrowRight/>}
-            </IconButton>
-            <IconButton
-                onClick={handleLastPageButtonClick}
-                disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                aria-label="last page"
-            >
-                {theme.direction === 'rtl' ? <FirstPageIcon/> : <LastPageIcon/>}
-            </IconButton>
-        </div>
-    );
-}
-
-TablePaginationActions.propTypes = {
-    count: PropTypes.number.isRequired,
-    onChangePage: PropTypes.func.isRequired,
-    page: PropTypes.number.isRequired,
-    rowsPerPage: PropTypes.number.isRequired,
-};
-
-
-function createData(sprite, name, weight, height, baseExperience) {
-    return {sprite, name, weight, height, baseExperience};
-}
-
-const rows = [
-    createData('https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/1.svg', 'bulbasaur', 50, 50, 50),
-    createData('https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/2.svg', 'bulbasaur1', 50, 50, 50),
-    createData('https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/3.svg', 'bulbasaur2', 50, 50, 50),
-    createData('https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/4.svg', 'bulbasaur3', 50, 50, 50),
-    createData('https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/5.svg', 'bulbasaur4', 50, 50, 50),
-    createData('https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/6.svg', 'bulbasaur5', 50, 50, 50),
-    createData('https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/7.svg', 'bulbasaur6', 50, 50, 50),
-]
-
-
-export class CustomPaginationActionsTable extends React.Component {
-    constructor() {
-        super();
-        this.state = {
-            page: 0,
-            rowsPerPage: 5
-        };
-        this.handleChangePage = this.handleChangePage.bind(this);
-        this.handleChangeRowsPerPage = this.handleChangeRowsPerPage.bind(this);
-    }
-
-    handleChangePage = (event, newPage) => {
-        this.setState({page: newPage});
-    };
-
-    handleChangeRowsPerPage = (event) => {
-        this.setState({rowsPerPage: parseInt(event.target.value, 10)});
-        this.setState({page: 0});
-    };
-
-    render() {
-        const emptyRows = this.state.rowsPerPage - Math.min(this.state.rowsPerPage, rows.length - this.state.page * this.state.rowsPerPage);
-        return (
-            <TableContainer component={Paper}>
-                <Table className="table" aria-label="custom pagination table">
-                    <TableHead>
-                        <TableRow>
-                            <TableCell>Sprite</TableCell>
-
-                            <TableCell>Name</TableCell>
-                            <TableCell>Weight</TableCell>
-                            <TableCell>Height</TableCell>
-                            <TableCell>Base experience</TableCell>
-
-                        </TableRow>
-                    </TableHead>
-                    <TableBody>
-                        {(this.state.rowsPerPage > 0
-                                ? rows.slice(this.state.page * this.state.rowsPerPage, this.state.page * this.state.rowsPerPage + this.state.rowsPerPage)
-                                : rows
-                        ).map((row) => (
-                            <TableRow key={row.sprite}>
-                                <TableCell style={{width: 160}} align="left">
-                                    <img src={row.sprite} style={{width: 100, height: 100}}/>
-                                </TableCell>
-                                <TableCell style={{width: 160}} align="left">
-                                    {row.name}
-                                </TableCell>
-                                <TableCell style={{width: 160}} align="left">
-                                    {row.weight}
-                                </TableCell>
-                                <TableCell style={{width: 160}} align="left">
-                                    {row.height}
-                                </TableCell>
-                                <TableCell style={{width: 160}} align="left">
-                                    {row.baseExperience}
-                                </TableCell>
-                            </TableRow>
-                        ))}
-
-                        {emptyRows > 0 && (
-                            <TableRow style={{height: 53 * emptyRows}}>
-                                <TableCell colSpan={6}/>
-                            </TableRow>
-                        )}
-                    </TableBody>
-                    <TableFooter>
-                        <TableRow>
-                            <TablePagination
-                                rowsPerPageOptions={[1, 5, 10, 25, {label: 'All', value: -1}]}
-                                colSpan={5}
-                                count={rows.length}
-                                rowsPerPage={this.state.rowsPerPage}
-                                page={this.state.page}
-                                SelectProps={{
-                                    inputProps: {'aria-label': 'Pokemons per page'},
-                                    native: true,
-                                }}
-                                onChangePage={this.handleChangePage}
-                                onChangeRowsPerPage={this.handleChangeRowsPerPage}
-                                ActionsComponent={TablePaginationActions}
-                            />
-                        </TableRow>
-                    </TableFooter>
-                </Table>
-            </TableContainer>
-        );
-    }
-}
