+++ src/components/table/table.component.style.css
@@ -1,0 +1,4 @@
+.root{
+    flex-shrink: 0;
+    margin-left: 2rem;
+}
